refactor: rename `idxslice` module to `slice`.
