+++ crates/oxc_index/src/slice.rs
